feat(dein): Updating plugins.vim to support Dein
Dein.vim calls and setup including the XDG base directory spec working
properly for the repository. Previously the suggestion was to use
Neobundle but it was discontinued in favor of Dein.vim, so here we are.

README and "doc" strings updated to reflect changes in the repository.
Fully working as expected ATM, replacing the previous Vundle
workarounds